cleanup functional test, throw error if missing 'java' plugin
