+++ starts-gradle-plugin/build.gradle
@@ -1,15 +1,7 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Gradle plugin project to get you started.
- * For more details take a look at the Writing Custom Plugins chapter in the Gradle
- * User Manual available at https://docs.gradle.org/7.5.1/userguide/custom_plugins.html
- */
-
 plugins {
-    id 'com.gradle.plugin-publish' version '1.0.0-rc-1'
     id 'java-gradle-plugin'
     id 'maven-publish'
+    id 'com.gradle.plugin-publish' version '1.0.0-rc-1'
 }
 
 project.description = 'STARTS Gradle Plugin'
@@ -29,6 +21,11 @@ dependencies {
     testImplementation 'org.junit.jupiter:junit-jupiter:5.8.2'
 }
 
+compileJava {
+    sourceCompatibility = 1.8
+    targetCompatibility = 1.8
+}
+
 pluginBundle {
     website = 'https://github.com/TestingResearchIllinois/starts'
     vcsUrl = 'https://github.com/TestingResearchIllinois/starts'
@@ -41,7 +38,7 @@ gradlePlugin {
             id = 'edu.illinois.starts'
             displayName = 'STARTS plugin for Gradle'
             description = 'STARTS plugin for Gradle'
-            implementationClass = 'edu.illinois.starts.gradle.plugin.StartsPlugin'
+            implementationClass = 'edu.illinois.starts.plugin.gradle.StartsPlugin'
         }
     }
 }
