--- starts-gradle-plugin/src/functionalTest/java/edu/illinois/starts/gradle/plugin/StartsPluginFunctionalTest.java
@@ -1,57 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package edu.illinois.starts.gradle.plugin;
-
-import org.gradle.testkit.runner.BuildResult;
-import org.gradle.testkit.runner.GradleRunner;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.Writer;
-
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-/**
- * A simple functional test for the 'edu.illinois.starts.gradle.plugin' plugin.
- */
-class StartsPluginFunctionalTest {
-    @TempDir
-    File projectDir;
-
-    private File getBuildFile() {
-        return new File(projectDir, "build.gradle");
-    }
-
-    private File getSettingsFile() {
-        return new File(projectDir, "settings.gradle");
-    }
-
-    @Test void canRunTask() throws IOException {
-        writeString(getSettingsFile(), "");
-        writeString(getBuildFile(),
-            "plugins {" +
-            "  id('edu.illinois.starts')" +
-            "}");
-
-        // Run the build
-        GradleRunner runner = GradleRunner.create();
-        runner.forwardOutput();
-        runner.withPluginClasspath();
-        runner.withArguments("startsHelp");
-        runner.withProjectDir(projectDir);
-        BuildResult result = runner.build();
-
-        // Verify the result
-        assertTrue(result.getOutput().contains("A tool for STAtic Regression Test Selection."));
-    }
-
-    private void writeString(File file, String string) throws IOException {
-        try (Writer writer = new FileWriter(file)) {
-            writer.write(string);
-        }
-    }
-}
