+++ starts-gradle-plugin/src/functionalTest/java/edu/illinois/starts/plugin/gradle/StartsPluginFunctionalTest.java
@@ -1,0 +1,68 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package edu.illinois.starts.plugin.gradle;
+
+import org.gradle.testkit.runner.BuildResult;
+import org.gradle.testkit.runner.GradleRunner;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
+
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.Writer;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+/**
+ * Functional tests for the 'edu.illinois.starts.plugin.gradle' plugin.
+ */
+class StartsPluginFunctionalTest {
+    @TempDir
+    static File projectDir;
+
+    GradleRunner runner;
+
+    private static File getBuildFile() {
+        return new File(projectDir, "build.gradle");
+    }
+
+    private static File getSettingsFile() {
+        return new File(projectDir, "settings.gradle");
+    }
+
+
+    private static void writeString(File file, String string) throws IOException {
+        try (Writer writer = new FileWriter(file)) {
+            writer.write(string);
+        }
+    }
+
+    @BeforeAll
+    public static void setupTest() throws IOException {
+        writeString(getSettingsFile(), "");
+        writeString(getBuildFile(), """
+                        plugins {
+                            id 'java'
+                            id 'edu.illinois.starts' version '1.4-SNAPSHOT'
+                        }""".stripIndent());
+    }
+
+    @BeforeEach
+    public void setupTests() {
+        runner = GradleRunner.create();
+        runner.forwardOutput();
+        runner.withPluginClasspath();
+        runner.withProjectDir(projectDir);
+    }
+
+    @Test
+    public void helpTask() {
+        runner.withArguments("startsHelp");
+        BuildResult result = runner.build();
+        assertTrue(result.getOutput().contains("A tool for STAtic Regression Test Selection."));
+    }
+}
