+++ starts-gradle-plugin/src/main/java/edu/illinois/starts/plugin/gradle/StartsPlugin.java
@@ -1,13 +1,13 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package edu.illinois.starts.gradle.plugin;
+package edu.illinois.starts.plugin.gradle;
 
-import edu.illinois.starts.gradle.plugin.tasks.CleanTask;
-import edu.illinois.starts.gradle.plugin.tasks.DiffTask;
-import edu.illinois.starts.gradle.plugin.tasks.HelpTask;
-import edu.illinois.starts.gradle.plugin.tasks.RunTask;
-import edu.illinois.starts.gradle.plugin.tasks.StartsTask;
+import edu.illinois.starts.plugin.gradle.tasks.CleanTask;
+import edu.illinois.starts.plugin.gradle.tasks.DiffTask;
+import edu.illinois.starts.plugin.gradle.tasks.HelpTask;
+import edu.illinois.starts.plugin.gradle.tasks.RunTask;
+import edu.illinois.starts.plugin.gradle.tasks.StartsTask;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
