+++ starts-gradle-plugin/src/main/java/edu/illinois/starts/plugin/gradle/tasks/CleanTask.java
@@ -1,4 +1,4 @@
-package edu.illinois.starts.gradle.plugin.tasks;
+package edu.illinois.starts.plugin.gradle.tasks;
 
 import edu.illinois.starts.helpers.FileUtil;
 import org.gradle.api.tasks.TaskAction;
