+++ starts-gradle-plugin/src/main/java/edu/illinois/starts/plugin/gradle/tasks/DiffTask.java
@@ -1,4 +1,4 @@
-package edu.illinois.starts.gradle.plugin.tasks;
+package edu.illinois.starts.plugin.gradle.tasks;
 
 import edu.illinois.starts.data.ZLCFormat;
 import edu.illinois.starts.enums.DependencyFormat;
