+++ starts-gradle-plugin/src/test/java/edu/illinois/starts/plugin/gradle/StartsPluginTest.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package edu.illinois.starts.gradle.plugin;
+package edu.illinois.starts.plugin.gradle;
 
 import org.gradle.testfixtures.ProjectBuilder;
 import org.gradle.api.Project;
@@ -22,5 +22,6 @@ class StartsPluginTest {
         assertNotNull(project.getTasks().findByName("startsClean"));
         assertNotNull(project.getTasks().findByName("startsDiff"));
         assertNotNull(project.getTasks().findByName("startsRun"));
+        assertNotNull(project.getTasks().findByName("starts"));
     }
 }
